+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/detail/VerseDetailPresenter.kt
@@ -37,11 +37,9 @@ import me.xizzhu.android.joshua.infra.arch.onEach
 import me.xizzhu.android.joshua.infra.arch.onEachSuccess
 import me.xizzhu.android.joshua.infra.interactors.BaseSettingsAwarePresenter
 import me.xizzhu.android.joshua.reading.ReadingActivity
-import me.xizzhu.android.joshua.reading.VerseDetailRequest
 import me.xizzhu.android.joshua.reading.verse.toStringForSharing
 import me.xizzhu.android.joshua.strongnumber.StrongNumberListActivity
 import me.xizzhu.android.joshua.ui.*
-import me.xizzhu.android.joshua.utils.supervisedAsync
 import me.xizzhu.android.logger.Log
 import java.lang.StringBuilder
 import kotlin.math.max
@@ -77,7 +75,10 @@ class VerseDetailPresenter(private val readingActivity: ReadingActivity,
         viewHolder.verseDetailViewLayout.post { viewHolder.verseDetailViewLayout.hide() }
 
         interactor.settings().onEachSuccess { viewHolder.verseDetailViewLayout.setSettings(it) }.launchIn(coroutineScope)
-        interactor.verseDetailRequest().onEach { showVerseDetail(it.verseIndex, it.content) }.launchIn(coroutineScope)
+        interactor.verseDetailRequest().onEach {
+            loadVerseDetail(it.verseIndex)
+            viewHolder.verseDetailViewLayout.show(it.content)
+        }.launchIn(coroutineScope)
         interactor.currentVerseIndex().onEach { close() }.launchIn(coroutineScope)
     }
 
@@ -190,24 +191,19 @@ class VerseDetailPresenter(private val readingActivity: ReadingActivity,
         downloadStrongNumberDialog = null
     }
 
-    private fun showVerseDetail(verseIndex: VerseIndex, @VerseDetailRequest.Companion.Content content: Int) {
-        loadVerseDetail(verseIndex)
-        viewHolder?.verseDetailViewLayout?.show(content)
-    }
-
     private fun loadVerseDetail(verseIndex: VerseIndex) {
         coroutineScope.launch {
             try {
                 viewHolder?.verseDetailViewLayout?.setVerseDetail(VerseDetail.INVALID)
-
-                val bookmarkAsync = supervisedAsync { interactor.readBookmark(verseIndex) }
-                val highlightAsync = supervisedAsync { interactor.readHighlight(verseIndex) }
-                val noteAsync = supervisedAsync { interactor.readNote(verseIndex) }
-                val strongNumberAsync = supervisedAsync { interactor.readStrongNumber(verseIndex) }
-
-                verseDetail = VerseDetail(verseIndex, buildVerseTextItems(verseIndex),
-                        bookmarkAsync.await().isValid(), highlightAsync.await().color,
-                        noteAsync.await().note, strongNumberAsync.await().toStrongNumberItems())
+                verseDetail = coroutineScope {
+                    val bookmarkAsync = async { interactor.readBookmark(verseIndex) }
+                    val highlightAsync = async { interactor.readHighlight(verseIndex) }
+                    val noteAsync = async { interactor.readNote(verseIndex) }
+                    val strongNumberAsync = async { interactor.readStrongNumber(verseIndex) }
+                    return@coroutineScope VerseDetail(verseIndex, buildVerseTextItems(verseIndex),
+                            bookmarkAsync.await().isValid(), highlightAsync.await().color,
+                            noteAsync.await().note, strongNumberAsync.await().toStrongNumberItems())
+                }
                 viewHolder?.verseDetailViewLayout?.setVerseDetail(verseDetail!!)
             } catch (e: Exception) {
                 Log.e(tag, "Failed to load verse detail", e)